refactor(ci-handler): extract module structure (P0-1 r1)
PR #484 r0's `src/mcp/handlers/ci.rs` reached 704 LOC after the new
contract test suite, tripping the `mcp_handler_files_under_500_loc`
file-size invariant (CI red on macos + ubuntu, MAX_LOC=700 in
tests/file_size_invariant.rs).

Mechanical refactor only — no logic change:
- Move impl + helpers (handle_repo_checkout, handle_release_repo,
  handle_watch_ci, handle_unwatch_ci, validate_release_path) to
  `src/mcp/handlers/ci/mod.rs`.
- Move all 17 tests (10 pre-existing + 7 P0-1 contract invariants) to
  `src/mcp/handlers/ci/tests.rs`, de-indented out of the `mod tests {}`
  wrapper. `use super::*` continues to resolve to the impl module.
- Same module-pattern precedent as `src/mcp/handlers/dispatch_hook/`
  (Sprint 53 P0-2 r1).

Verification:
- `cargo test --test file_size_invariant`: 1 passed (PR #484 unblocked).
- `cargo test --bin agend-terminal mcp::handlers::ci`: 17 passed.
- `cargo clippy --all-targets -- -D warnings`: clean.
- Full bin test suite: 1556 passed + 7 pre-existing flakes (unchanged
  from r0; documented as Sprint 54 P2-7 candidates).


Agend-Agent: dev
Agend-Branch: dev/sprint-54-p0-1-ci-watch-split
Agend-Issued-At: 2026-05-07T00:32:04.412487+00:00
